requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
python-multipart
jinja2
requests